feat: types refactoring 1) Move all .d.ts files, which provide types for different packages, to packages folder. 2) Rename interfaces.ts to restaurants.ts, because this file will contain types related to restaurants data.
